Add ASRSnippet, Target and Campaign models.
