+++ snippets/base/admin/actions.py
@@ -1,0 +1,8 @@
+from django.db import transaction
+
+
+@transaction.atomic
+def duplicate_snippets_action(modeladmin, request, queryset):
+    for snippet in queryset:
+        snippet.duplicate()
+duplicate_snippets_action.short_description = 'Duplicate selected snippets'  # noqa
